requirements.txt
cycler==0.11.0
essential-generators==1.0
fonttools==4.32.0
imageio==2.16.2
imgaug==0.4.0
kiwisolver==1.4.2
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pycairo==1.21.0
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.2
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
tifffile==2022.4.8
